// projects/services/src/lib/domain/models/customer-model.ts
export interface CustomerModel {
        customerID?: number;
        firstName?: string;
        lastName?: string;
        email?: string;
        phone?: string;
        address?: string;
        city?: string;
        country?: string;
        base64Url?: string;
        urlImage?: string;
        password?: string;
        acceptTerms?: boolean;
}